+++ model3d.py
@@ -1,0 +1,89 @@
+import os
+
+fname = os.path.join(os.path.dirname(__file__), "../meshes/foo3d")
+vname = os.path.join(os.path.dirname(__file__), "../velocity_models/foo3d")
+
+# Define mesh file to be used:
+meshfile = fname
+
+# Define initial velocity model:
+truemodel = vname
+initmodel = vname
+
+
+# Choose method and parameters
+opts = {
+    "method": "CG",
+    "quadrature": "KMV",
+    "variant": None,
+    "type": "SIP",  # for DG only - SIP, NIP and IIP
+    "degree": 1,  # p order
+    "dimension": 3,  # dimension
+    "mesh_size": 0.005,  # h
+    "beta": 0.0,  # for Newmark only
+    "gamma": 0.5,  # for Newmark only
+}
+
+parallelism = {
+    "type": "automatic",  # options: automatic, custom, off
+    "custom_cores_per_shot": [],  # only if the user wants a different number of cores for every shot.
+    # input is a list of integers with the length of the number of shots.
+}
+
+mesh = {
+    "Lz": 2.000,  # depth in km - always positive
+    "Lx": 1.0,  # width in km - always positive
+    "Ly": 1.0,  # thickness in km - always positive
+    "meshfile": meshfile + ".msh",
+    "initmodel": initmodel + ".hdf5",
+    "truemodel": truemodel + ".hdf5",
+}
+
+BCs = {
+    "status": False,  # True or false
+    "outer_bc": "non-reflective",  # Neuumann, non-reflective (outer boundary condition)
+    "damping_type": "polynomial",  # polynomial. hyperbolic, shifted_hyperbolic
+    "exponent": 1,
+    "cmax": 4.7,  # maximum acoustic wave velocity in PML - km/s
+    "R": 0.001,  # theoretical reflection coefficient
+    "lz": 0.0,  # thickness of the pml in the z-direction (km) - always positive
+    "lx": 0.0,  # thickness of the pml in the x-direction (km) - always positive
+    "ly": 0.0,  # thickness of the pml in the y-direction (km) - always positive
+}
+
+acquisition = {
+    "source_type": "Ricker",
+    "frequency": 2.0,
+    "delay": 1.0,
+    "source_pos": [(-0.5, 0.5, 0.5)],  # z,x,y
+    "receiver_locations": [()],
+}  # equi-spaced for now
+
+timeaxis = {
+    "t0": 0.0,  # Initial time for event
+    "tf": 0.20,  # Final time for event
+    "dt": 0.001,  # timestep size
+    "nspool": 20,  # how frequently to output solution to pvds
+    "fspool": 10,  # how frequently to save solution to RAM
+    "amplitude": 1,  # the Ricker has an amplitude of 1.
+}  # how freq. to output to files and screen
+
+inversion = {
+    "freq_bands": [None]
+}  # cutoff frequencies (Hz) for Ricker source and to low-pass the observed shot record
+
+aut_dif = {
+    "status": False,
+}
+# Create your model with all the options
+model = {
+    "self": None,
+    "inversion": inversion,
+    "opts": opts,
+    "BCs": BCs,
+    "parallelism": parallelism,
+    "mesh": mesh,
+    "acquisition": acquisition,
+    "timeaxis": timeaxis,
+    "aut_dif": aut_dif,
+}
